+++ src/Identity/Logitar.Identity.Domain/Roles/Events/RoleDeletedEvent.cs
@@ -1,0 +1,12 @@
+﻿using Logitar.EventSourcing;
+using MediatR;
+
+namespace Logitar.Identity.Domain.Roles.Events;
+
+public record RoleDeletedEvent : DomainEvent, INotification
+{
+  public RoleDeletedEvent()
+  {
+    DeleteAction = DeleteAction.Delete;
+  }
+}