+++ src/Identity/Logitar.Identity.Domain/Settings/RoleSettings.cs
@@ -1,0 +1,10 @@
+﻿namespace Logitar.Identity.Domain.Settings;
+
+public record RoleSettings : IRoleSettings
+{
+
+  public IUniqueNameSettings UniqueNameSettings { get; set; } = new UniqueNameSettings
+  {
+    AllowedCharacters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-_"
+  };
+}